Updating the changes from git. Changes are added seperate class for the encryption function in login controller
