+++ AddressBook/AddressBook.csproj
@@ -136,6 +136,8 @@
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
+    <Compile Include="ViewModels\Countries.cs" />
+    <Compile Include="Helpers\CustomHashs.cs" />
     <Compile Include="Models\AddressBookModel.Context.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
